resetboard seems to solve all of my problems with crashes.
Arduino Client (SolidSolis Firmata) and Arduino Board have to be in sync.
I have an issue where I reconnect to a board after closing and it gives F8 byte error or smth because the board is still sending analog firmata signals (like temperature change) and I want to establish a new connection already. This fix is the only way, otherwise I will have to dig through the SolidSolis code or switch to C++ (which does restart the board every time I connect to it via comm port).